Conflicts:
	Sources/Internal/MemoryManager/MemoryManager.cpp
	Sources/Internal/MemoryManager/MemoryManager.h
	Sources/Internal/MemoryManager/MemoryManagerTypes.h
	Sources/Internal/MemoryManager/MemoryProfiler.h
	Tools/ResourceEditor/Classes/Qt/DeviceInfo/DeviceInfo/MemProfController.cpp
	Tools/ResourceEditor/Classes/Qt/DeviceInfo/DeviceInfo/MemoryItemStyleDelegate.cpp

Former-commit-id: 03a5d7077676afc29567206fdc20010f5a0920fb

+++ Sources/Internal/MemoryManager/AllocThunk.cpp
@@ -35,7 +35,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 namespace DAVA
 {
 
-void* AllocThunk(size_t size, unsigned int poolIndex)
+void* AllocThunk(size_t size, size_t poolIndex)
 {
     return MemoryManager::Instance()->Allocate(size, poolIndex);
 }
